+++ src/main/java/domain/hint/Ball.java
@@ -1,4 +1,4 @@
-package domain;
+package domain.hint;
 
 public class Ball {
