+++ src/main/java/domain/hint/Strike.java
@@ -1,4 +1,4 @@
-package domain;
+package domain.hint;
 
 public class Strike {
