+++ src/main/java/domain/number/ComputerBallNumber.java
@@ -1,4 +1,6 @@
-package domain;
+package domain.number;
+
+import domain.generator.BaseBallNumberGenerator;
 
 public class ComputerBallNumber {
